refactor `pr list` to use same filter flags as issues
previously we used the /pulls endpoint, but that didn't support many
filters (https://gitea.com/api/swagger#/repository/repoListPullRequests).
Instead, we now do an issue listing, which also supports listing PRs.
The drawback here is that it only returns issue fields, but none of a PR's
special fields like head repo, base branch, mergeable flag, ...
This is a limitation of the upstream API, and we should carefully decide which
limitation to take, as changes to the API etc will take quite some time to propagate
through the ecosystem

This commit also adds the --fields flag to `tea pulls list`, with the
same fields available as issues (see above)